Re-organise widgets and handlers
