+++ test/app_testing/test_apps/tb_streaming/config/config_fed_server.json
@@ -19,7 +19,7 @@
     },
     {
       "id": "event_recorder",
-      "path": "nvflare.app_common.handlers.event_recorder.ServerEventRecorder",
+      "name": "nvflare.widgets.event_recorder.ServerEventRecorder",
       "args": {}
     },
     {
